// src/statistics/enums/statistics-fields-name.enum.ts
export enum StatisticsFieldsName {
	NUMBER_WORK_DAYS = 'numberWorkDays',
	NUMBER_DAYS_OFF = 'numberDaysOff',
	NUMBER_VACATION_DAYS = 'numberVacationDays',
	NUMBER_SICK_DAYS = 'numberSickDays',
	NUMBER_ADDITIONAL_WORK_DAYS = 'numberAdditionalWorkDays',
	NUMBER_WORK_HOURS = 'numberWorkHours',
	NUMBER_FREE_HOURS = 'numberFreeHours',
	NUMBER_VACATION_HOURS = 'numberVacationHours',
	NUMBER_SICK_HOURS = 'numberSickHours',
	NUMBER_ADDITIONAL_WORK_HOURS = 'numberAdditionalWorkHours',
	TOTAL_DAYS = 'totalDays',
	TOTAL_HOURS = 'totalHours',
	NUMBER_FIRST_SHIFTS = 'numberFirstShifts',
	NUMBER_SECOND_SHIFT = 'numberSecondShifts',
	NUMBER_NIGHT_HOURS = 'numberNightHours',
	GROSS_AMOUNT_MONEY_FOR_WORK_DAYS = 'grossAmountMoneyForWorkDays',
	NET_AMOUNT_MONEY_FOR_WORK_DAYS = 'netAmountMoneyForWorkDays',
	GROSS_AMOUNT_MONEY_FOR_VACATION_DAYS = 'grossAmountMoneyForVacationDays',
	NET_AMOUNT_MONEY_FOR_VACATION_DAYS = 'netAmountMoneyForVacationDays',
	GROSS_AMOUNT_MONEY_FOR_SICK_DAYS = 'grossAmountMoneyForSickDays',
	NET_AMOUNT_MONEY_FOR_SICK_DAYS = 'netAmountMoneyForSickDays',
	TOTAL_MONEY_EARNED_GROSS = 'totalMoneyEarnedGross',
	TOTAL_MONEY_EARNED_NET = 'totalMoneyEarnedNet',
	TOTAL_TAX_PAID = 'totalTaxPaid',
}